Update makeuser
added banned user support and now moves added users to signups_current file instead of having to go through the  large list every time new users are added.    Also touches the .new_user file to ensure the date is the date the user was created so it'll be easier to figure out stale new user accounts in future.